6/16(金) ・宮城の情報統合 ・富永の情報統合 6/17(土) ・内間の情報統合 ・親富祖の情報統合 6/18(日) ・コミュニティに再参加する場合は、community_userテーブルをINSERTではなく、UPDATEする。 ・ニックネームを付けない場合は、ユーザー名で参加する ・お知らせ一覧のタブ作成 ・不要なファイルの削除 ・プロフィール閲覧機能統合 ・ファイル統合
