--- src/main/java/com/example/Kuchannel/dao/UchimaDao.java
@@ -1,69 +1,0 @@
-package com.example.Kuchannel.dao;
-
-import com.example.Kuchannel.entity.BelongingCommunities;
-import com.example.Kuchannel.entity.MyReview;
-import com.example.Kuchannel.entity.MyThread;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.jdbc.core.DataClassRowMapper;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public class UchimaDao {
-
-    @Autowired
-    private NamedParameterJdbcTemplate jdbcTemplate;
-
-    //マイページ用で、所属しているコミュニティを取得する処理
-    public List<BelongingCommunities> getBelongingCommunities(Integer userId){
-        MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("userId", userId);
-        var result = jdbcTemplate.query("SELECT c.id AS communityId ,c_u.user_id,c.name AS community_name,c_u.nick_name ,c_u.flag FROM community_user c_u JOIN communities c ON c_u.community_id = c.id WHERE c_u.user_id = :userId",param,new DataClassRowMapper<>(BelongingCommunities.class));
-        System.out.println(result);
-        return result;
-    }
-
-    //マイページ用で、コミュニティから退会する処理
-    public int withdrawal(Integer userId,Integer communityId){
-        MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("userId", userId);
-        param.addValue("communityId", communityId);
-        var result = jdbcTemplate.update("UPDATE community_user SET flag = false WHERE user_id = :userId AND community_id = :communityId",param);
-        return result;
-    }
-
-    //マイページ用で、コミュニティのニックネームを変更する処理
-    public int updateNickName(BelongingCommunities updateInfo){
-        MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("nickName",   updateInfo.getNickName());
-        param.addValue("userId", updateInfo.getUserId());
-        param.addValue("communityId", updateInfo.getCommunityId());
-        var result = jdbcTemplate.update("UPDATE community_user SET nick_name = :nickName WHERE user_id = :userId AND community_id = :communityId",param);
-        return result;
-    }
-
-//    マイページ用で、自分が建てたスレッドを取得する処理
-    public List<MyThread>getMyThreads(Integer userId){
-        MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("userId", userId);
-        var result = jdbcTemplate.query("SELECT th.id AS threadId, th.title AS threadTitle, co.name AS communityName, co.url AS communityUrl " +
-                "FROM threads th JOIN communities co ON th.community_id = co.id JOIN community_user cu ON co.id = cu.community_id " +
-                "WHERE th.user_id = :userId AND cu.user_id = :userId AND cu.flag = true;", param,
-                new DataClassRowMapper<>(MyThread.class));
-        System.out.println(result);
-        return result;
-    }
-
-    //    マイページ用で、自分で書いたレビュー取得する処理
-    public List<MyReview>getMyReviews(Integer userId){
-        MapSqlParameterSource param = new MapSqlParameterSource();
-        param.addValue("userId", userId);
-        var result = jdbcTemplate.query("SELECT rev.title AS reviewTitle, rev.review, th.title AS threadTitle, co.name AS communityName, rev.create_date AS createDate FROM reviews rev JOIN threads th ON rev.thread_id = th.id JOIN communities co ON th.community_id = co.id JOIN community_user cu on co.id = cu.community_id WHERE th.user_id = :userId AND cu.flag = true;",param,new DataClassRowMapper<>(MyReview.class));
-        System.out.println(result);
-        return result;
-    }
-
-}
