--- src/main/resources/templates/profile.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>プロフィール画面</title>
-</head>
-<body>
-
-</body>
-</html>
